+++ src/hybm/csrc/ts_engine/hybm_stream.cpp
@@ -110,7 +110,7 @@ int32_t HybmStream::AllocLogicCq()
     input.cqId = 65535U;
     input.sqId = 0;
     input.info[0] = streamId_;
-    input.info[1] = static_cast<uint32_t>(syscall(SYS_gettid));
+    input.info[1] = syscall(SYS_gettid);
 
     auto ret = DlHalApi::HalSqCqAllocate(deviceId_, &input, &output);
     if (ret != 0) {
